% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{aggregate_csv}
\alias{aggregate_csv}
\title{Find and combine multiple csv files into 1 xlsx file}
\usage{
aggregate_csv(path)
}
\arguments{
\item{path}{The path where the csv's will be searched and output will be generated}
}
\description{
Searches path recursively for csv files. If csv files are
found in a subdir, an aggregated xlsx file of them is generated, with the
csv filenames as tabnames of the xlsx file.
}
\examples{
example_output_directory <- paste0(tempdir(), "/csv_files")
fs::dir_create(example_output_directory)
readr::write_csv(cars, fs::path(example_output_directory, "cars1.csv"))
readr::write_csv(cars, fs::path(example_output_directory, "cars2.csv"))
aggregate_csv(example_output_directory)
system(paste0("ls ", example_output_directory))
}
