% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio.R
\name{annotate_results_table}
\alias{annotate_results_table}
\title{Annotates deseq result with metadata from a SummarizedExperiment}
\usage{
annotate_results_table(deseq_results, se, row_data_cols_to_add)
}
\arguments{
\item{deseq_results}{A DESeq results table returned from DESeq2::results()}

\item{se}{The SummarizedExperiment from which to extract the additional data.
Must contain a column named ensg in its rowdata}

\item{row_data_cols_to_add}{a character vector containing the colnames from
the se's rowdata}
}
\value{
a tibble with the information added and incomplete cases removed
}
\description{
The joining will be done on the ensg column
}
