% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio.R
\name{automap_ids}
\alias{automap_ids}
\title{Detects annotation identifiers and converts them to the requested format}
\usage{
automap_ids(ids, return_keytype = "SYMBOL", species = "human")
}
\arguments{
\item{ids}{vector containing the original ids, all of the same keytype}

\item{return_keytype}{format which to convert ids to. Options can be found
by running library(org.Hs.eg.db); keytypes(org.Hs.eg.db)}

\item{species}{either "human" or "mouse" to indicate from which species the
identifiers are}
}
\value{
named vector containing the converted ids, named with the original
ids
}
\description{
Uses the org.Hs.eg.db database to detect the keytype of the input vector
and converts it to the keytype defined by return_keytype.
When multiple target ids are found for a source id, the first will be
returned.
}
\examples{
ensembl_ids <- c("ENSG00000132475.8", "ENSG00000185339.8")
symbols <- automap_ids(ensembl_ids)
}
