% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio.R
\name{calculate_contrasts_table}
\alias{calculate_contrasts_table}
\title{Creates a contrasts table}
\usage{
calculate_contrasts_table(se, contrast, design_char, ...)
}
\arguments{
\item{se}{A SummarizedExperiment containing the expression data and the
metadata required in the defined contrast}

\item{contrast}{The variables to contrast. See ?DESeq2::results}

\item{design_char}{a character that will be converted to the design formula.
See ?DESeq2::DESeqDataSet}

\item{...}{additional arguments passed to DESeq2::results}
}
\value{
A tibble with the requested contrast table
}
\description{
The joining will be done on the ensg column
}
