% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{colored_dot_plot}
\alias{colored_dot_plot}
\title{Creates dot plot using defined columns colored by color_col}
\usage{
colored_dot_plot(data, x_col, y_col, color_col)
}
\arguments{
\item{data}{A data frame or tibble in long format containing the columns
specified by x_col, y_col, and color_col}

\item{x_col}{The column which will be passed as the aes x arg to ggplot}

\item{y_col}{The column which will be passed as the aes y arg to ggplot}

\item{color_col}{The column that will be used to color the points}
}
\value{
The plot
}
\description{
Creates dot plot using defined columns colored by color_col
}
\examples{
# plots disp vs mpg colored by gear for mtcars data
colored_dot_plot(mtcars, "disp", "mpg", "gear")
}
