% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contrasts_table}
\alias{contrasts_table}
\title{An example contrasts table generated by the create_contrasts_table function}
\format{
A tibble with 50 rows and 9 columns
\describe{
  \item{ensembl_id}{A dummy ensembl id}
  \item{log2FoldChange}{The log2 fold change of the gene}
  \item{pvalue}{The p value of the comparison}
  \item{padj}{The p value of the comparison corrected for multiple testing}
  \item{symbol}{The dummy HUGO symbol corresponding to the ensembl ID}
  \item{biotype}{The biotype of the gene}
  ...
}
}
\usage{
contrasts_table
}
\description{
The table contains a set of 50 random genes of which half are relevant
(low p value and high log2FoldChange)
}
\keyword{datasets}
