% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_empty_dir}
\alias{create_empty_dir}
\title{Creates an empty directory}
\usage{
create_empty_dir(path, mindepth = 5)
}
\arguments{
\item{path}{The resulting empty directory. If not existing, dir will be created.
!! If existing, dir will be emptied. !!}

\item{mindepth}{Depth of output directory. Used as safety check to avoid emptying high-level dirs}
}
\value{
The path to the directory
}
\description{
Creates an empty directory
}
\examples{
# New directory
directory <- paste0(tempdir(), "/example_creation_dir")
create_empty_dir(directory, 1)
# Existing non-empty directory (first example continues)
system(paste0("touch ", directory, "/newfile"))
create_empty_dir(directory, 1)
# -> newfile is deleted
}
