% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{csv_combine_from_path}
\alias{csv_combine_from_path}
\title{Combine multiple csv files into 1 xlsx file}
\usage{
csv_combine_from_path(path, recurse = TRUE)
}
\arguments{
\item{path}{The path where the csv's will be searched and output will be generated}

\item{recurse}{If TRUE, search for and combine csv files recursively from path}
}
\description{
Searches path for csv files. If csv files are found , an aggregated xlsx file
of them is generated, with the csv filenames as tabnames of the xlsx file.
}
\examples{
example_output_directory <- paste0(tempdir(), "/csv_files")
fs::dir_create(example_output_directory)
readr::write_csv(cars, fs::path(example_output_directory, "cars1.csv"))
readr::write_csv(cars, fs::path(example_output_directory, "cars2.csv"))
csv_combine_from_path(example_output_directory)
system(paste0("ls ", example_output_directory))
}
