% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{ear}
\alias{ear}
\title{Selects the top-left corner of a 2d data object}
\usage{
ear(x, n = 10)
}
\arguments{
\item{x}{A 2-dimensional data object such as data frame, tibble, or matrix}

\item{n}{Top n rows and columns to select}
}
\value{
The subsetted object
}
\description{
An extension of the head() function, which also selects the n leftmost
columns after applying head().
}
\examples{
ear(mtcars, 5)
}
