% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{joint_legends_grid_plot}
\alias{joint_legends_grid_plot}
\title{Creates a grid from plots with a common legend and only prints the legend once}
\usage{
joint_legends_grid_plot(plotlist, ...)
}
\arguments{
\item{plotlist}{list of ggplot objects}

\item{...}{additional arguments to plot_grid (left panel)}
}
\value{
The plot grid
}
\description{
Note that no check for equal legends is performed! The legend of the first
plot will be used.
}
\examples{
p1 <- p2 <- colored_dot_plot(mtcars, "disp", "mpg", "gear")
joint_legends_grid_plot(list(p1,p2))
}
