% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{pdf_combine_from_path}
\alias{pdf_combine_from_path}
\title{Combine multiple pdfs into 1 file}
\usage{
pdf_combine_from_path(path)
}
\arguments{
\item{path}{The path where the pdfs will be searched and output will be generated}
}
\description{
Finds all pdf files in path and combines them into one file with the name
}
