% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_nested_input}
\alias{read_nested_input}
\title{Reads RDS object from nested directories and creates a nested list from them}
\usage{
read_nested_input(dir_in)
}
\arguments{
\item{dir_in}{The directory where the input RDS files are located (character or fs::path)}
}
\value{
The nested list containing the content of the RDS files
}
\description{
Acts as a complement to write_nested_output(). The RDS files written to disk
by write_nested_output() will be read in to recreate the list originally
used by write_nested_output() to write the data to disk.
}
\examples{
# write output by running the example from write_any_output(), then run ...
\dontrun{nested_input <- read_nested_input(dir_out)}
}
