% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{repeat_over_args}
\alias{repeat_over_args}
\title{Repeatedly applies function .f on .data using args consecutively}
\usage{
repeat_over_args(.data, .f, args)
}
\arguments{
\item{.data}{An object to which the function will be applied}

\item{.f}{The function that will be applied}

\item{args}{An iterable of arguments that will be passed to the function}
}
\value{
The data after applying the function repeatedly
}
\description{
Repeatedly applies function .f on .data using args consecutively
}
\examples{
# adds 1 and then 2 to the vector
repeat_over_args(1:5, `+`, c(1,2))
}
