% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compbio.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Creates a volcano plot from a contrasts table}
\usage{
volcano_plot(
  contrast_tb,
  ttl = NULL,
  relevant_fc = 8,
  significance_cutoff = 0.01,
  add_labels = TRUE
)
}
\arguments{
\item{contrast_tb}{A tibble produced by facilda::calculate_contrasts_table}

\item{ttl}{The title of the plot}

\item{relevant_fc}{the absolute fold change above which genes are considered
relevant}

\item{significance_cutoff}{the adjusted p-value below which genes are
considered significant}

\item{add_labels}{whether to label the relevant points}
}
\value{
A ggplot2 object of the volcano plot
}
\description{
Genes for which the absolute fold change is larger than indidated and have
have an adjusted p value below the significance_cutoff will be highlighted
and labeled (when requested).
}
