% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_nested_output}
\alias{write_nested_output}
\title{Writes writable objects from a (nested) list}
\usage{
write_nested_output(output_list, dir_out, mindepth = 5)
}
\arguments{
\item{output_list}{A nested list}

\item{dir_out}{The directory where the output will be written to (character or fs::path)}

\item{mindepth}{Depth of output directory. Used as safety check to avoid emptying high-level dirs}
}
\value{
The path where the output was written to
}
\description{
Descends into a (nested) list until it finds either a data.frame or ggplot
object. These objects are then written as csv or A4 format pdf respectively,
Using the names of the list as subsequent folder and lastly file names.
Typically used to gather different outputs in a list which can then be
written.
NOTE: !! dir_out will be overwritten !!
}
\examples{
library(ggplot2)
cars_df <- cars
cars_plot <- ggplot(cars, aes(x=speed, y=dist)) + geom_point()
my_output_list <- list(my_output = list(plots = list(cars = cars_plot),
                                        tables = list(cars = cars_df)))
example_output_directory <- paste0(tempdir(), "/example_output")
dir_out <- write_nested_output(my_output_list, example_output_directory, 1)
# Structure of output list
str(my_output_list, max.level = 3)
# Structure in the returned temporary folder, e.g.
system(paste0("tree ", dir_out))
}
